+++ demo/SocialLogin/androidSDK/src/main/java/com/loginradius/sdk/models/page/LoginRadiusCountryCodeName.java
@@ -1,0 +1,10 @@
+package com.loginradius.sdk.models.page;
+import com.google.gson.annotations.SerializedName;
+
+public class LoginRadiusCountryCodeName 
+{
+	@SerializedName("Code")
+	public String Code ;
+	@SerializedName("Name")
+    public String Name ;
+}