+++ demo/SocialLogin/androidSDK/src/main/java/com/loginradius/sdk/util/AsyncHandler.java
@@ -1,0 +1,6 @@
+package com.loginradius.sdk.util;
+
+public interface AsyncHandler<T> {
+	public void onSuccess(T data);
+	public void onFailure(Throwable error, String errorcode);
+}