+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 > **LoginRadius Android SDK Change Log** provides information regarding what has changed, more specifically what changes, improvements and bug fix has been made to the SDK. For more details please refer to the [LoginRadius API Documention](http://apidocs.loginradius.com/docs/android)
 
 ### Version 2.0.0
-Released on **July 18,  2016**
+Released on **September 15,  2016**
 
 ##### Enhancements
 
@@ -19,3 +19,5 @@ Released on **July 18,  2016**
   
 ##### Bug Fixes
 - Handled social ID provider Empty email address scenario. 
+- Fix When back in App crash SDK Activity issue.
+- Fix Google Native Login Crash issue if the users device does not have Google Playservices installed.